Track contract JSON as dependency
This instructs webpack to reload when the compiled JSON file is modified. Otherwise, only the solidity file is tracked. This causes the artifact *not* to be reloaded when the user creates a new instance (since the JSON file is modified with the deployed address, but the solidity file is untouched), at least when the hot-loader is configured but disabled (eg in tutorial-kit).